Rename folder to hooks
